# Parallelize initialization of storageDisks