## Description
Parallelize initialization of storageDisks

## Motivation and Context
A higher number of endpoints should be initialized in parallel 

## How to test this PR?
```
~  benchcmp /tmp/prev /tmp/new 
benchmark                           old ns/op     new ns/op     delta
BenchmarkInitStorageDisks256-4      6868523       5633617       -17.98%
BenchmarkInitStorageDisks1024-4     31059490      26246112      -15.50%
BenchmarkInitStorageDisks2048-4     62277295      38179426      -38.69%
BenchmarkInitStorageDisksMax-4      193600127     178121000     -8.00%
```

```
~ benchstat /tmp/prev /tmp/new 
name                    old time/op    new time/op    delta
InitStorageDisks256-4     6.87ms ± 0%    5.63ms ± 0%   ~     (p=1.000 n=1+1)
InitStorageDisks1024-4    31.1ms ± 0%    26.2ms ± 0%   ~     (p=1.000 n=1+1)
InitStorageDisks2048-4    62.3ms ± 0%    38.2ms ± 0%   ~     (p=1.000 n=1+1)
InitStorageDisksMax-4      194ms ± 0%     178ms ± 0%   ~     (p=1.000 n=1+1)
```

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [ ] Fixes a regression (If yes, please add `commit-id` or `PR #` here)
- [ ] Documentation needed
- [x] Unit tests needed
- [ ] Functional tests needed (If yes, add [mint](https://github.com/minio/mint) PR # here: )