@minio-ops 2019-09-23
## Mint Automation
| Test | Result |
|:--- |:--- |
| mint-compression-xl.sh | :heavy_check_mark: |
| mint-compression-dist-xl.sh | :heavy_check_mark: |
| mint-compression-fs.sh | :heavy_check_mark: |
| mint-worm.sh | :heavy_check_mark: |
| mint-dist-xl.sh | :heavy_check_mark: |
| mint-gateway-nas.sh | :heavy_check_mark: |
| mint-xl.sh | :x: [more...](#daa52e74-f81b-4aa3-b03d-3e611468061a) |
| mint-large-bucket.sh | :x: [more...](#91e8b566-43c4-4bd9-bb96-73b95c288a98) |
| mint-fs.sh | :x: [more...](#5eb00a3b-5e05-475a-b14a-d9bdfac7203c) |
<a name="daa52e74-f81b-4aa3-b03d-3e611468061a"></a>
### 8288-306d2c0/mint-xl.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.58:32528
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 92524dbb68b7:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... FAILED in 6 minutes and 5 seconds
{
  "alert": "",
  "args": {
    "bucketName": "aws-sdk-go-test-r0jojengrusttu",
    "expiry": 60000000000,
    "objectName": "presignedTest"
  },
  "duration": 364888,
  "error": "XMinioServerNotInitialized: Server not initialized, please try again.\n\tstatus code: 503, request id: 15C6EE70B34A0DF8, host id: ",
  "function": "PresignedPut",
  "message": "AWS SDK Go CreateBucket Failed",
  "name": "aws-sdk-go",
  "status": "FAIL"
}

Executed 0 out of 14 tests successfully.
```
<a name="5eb00a3b-5e05-475a-b14a-d9bdfac7203c"></a>
### 8288-306d2c0/mint-fs.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.61:30150
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 50ddf311dc94:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... done in 0 seconds
(2/14) Running aws-sdk-java tests ... done in 1 seconds
(3/14) Running aws-sdk-php tests ... done in 42 seconds
(4/14) Running aws-sdk-ruby tests ... done in 2 seconds
(5/14) Running awscli tests ... done in 1 minutes and 52 seconds
(6/14) Running healthcheck tests ... done in 0 seconds
(7/14) Running mc tests ... done in 42 seconds
(8/14) Running minio-dotnet tests ... done in 25 seconds
(9/14) Running minio-go tests ... done in 53 seconds
(10/14) Running minio-java tests ... done in 51 seconds
(11/14) Running minio-js tests ... done in 47 seconds
(12/14) Running minio-py tests ... FAILED in 2 minutes and 10 seconds
{
  "duration": 54,
  "function": "presigned_get_object(bucket_name, object_name, expires, response_headers, request_date)",
  "args": {
    "response.status": 403,
    "response.reason": "Forbidden",
    "response._body": "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<Error><Code>AccessDenied</Code><Message>Request has expired</Message><Key>2aab3491-0d9d-4074-ab9d-6bbe42af54e9</Key><BucketName>minio-py-test-b275a1a9-8a88-4b9a-a611-b43bbfb71a69</BucketName><Resource>/minio-py-test-b275a1a9-8a88-4b9a-a611-b43bbfb71a69/2aab3491-0d9d-4074-ab9d-6bbe42af54e9</Resource><RequestId>15C6EE92A4114580</RequestId><HostId>f171c762-c63a-419f-8d1a-ea8bfc47e2a1</HostId></Error>",
    "object_name": "2aab3491-0d9d-4074-ab9d-6bbe42af54e9",
    "bucket_name": "minio-py-test-b275a1a9-8a88-4b9a-a611-b43bbfb71a69",
    "presigned_get_object_url": "http://72.28.97.61:30150/minio-py-test-b275a1a9-8a88-4b9a-a611-b43bbfb71a69/2aab3491-0d9d-4074-ab9d-6bbe42af54e9?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=minio%2F20190923%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20190923T015707Z&X-Amz-Expires=5&X-Amz-SignedHeaders=host&X-Amz-Signature=8179c6bf94b3fcdc403eb8d7a6f1ff8270afa82a743159d8174fcd85756400ae",
    "response.headers": "{\"_container\": {\"accept-ranges\": [\"Accept-Ranges\", \"bytes\"], \"content-length\": [\"Content-Length\", \"440\"], \"content-security-policy\": [\"Content-Security-Policy\", \"block-all-mixed-content\"], \"content-type\": [\"Content-Type\", \"application/xml\"], \"server\": [\"Server\", \"MinIO/DEVELOPMENT.2019-09-23T01-35-44Z\"], \"vary\": [\"Vary\", \"Origin\"], \"x-amz-request-id\": [\"X-Amz-Request-Id\", \"15C6EE92A4114580\"], \"x-xss-protection\": [\"X-Xss-Protection\", \"1; mode=block\"], \"date\": [\"Date\", \"Mon, 23 Sep 2019 01:57:13 GMT\"]}}"
  },
  "message": "AccessDenied: message: Access Denied",
  "name": "minio-py:test_presigned_get_object_expiry_5sec",
  "status": "FAIL",
  "error": "Traceback (most recent call last):\n  File \"/mint/run/core/minio-py/tests.py\", line 1341, in test_presigned_get_object_expiry_5sec\n    object_name).get_exception()\nminio.error.AccessDenied: AccessDenied: message: Access Denied\n\nDuring handling of the above exception, another exception occurred:\n\nTraceback (most recent call last):\n  File \"/mint/run/core/minio-py/tests.py\", line 2036, in main\n    test_presigned_get_object_expiry_5sec(client, log_output)\n  File \"/mint/run/core/minio-py/tests.py\", line 1355, in test_presigned_get_object_expiry_5sec\n    raise Exception(err)\nException: AccessDenied: message: Access Denied\n"
}

Executed 11 out of 14 tests successfully.
```
<a name="91e8b566-43c4-4bd9-bb96-73b95c288a98"></a>
### 8288-306d2c0/mint-large-bucket.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.55:31936
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 296037a77710:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... done in 3 seconds
(2/14) Running aws-sdk-java tests ... done in 5 seconds
(3/14) Running aws-sdk-php tests ... done in 53 seconds
(4/14) Running aws-sdk-ruby tests ... done in 15 seconds
(5/14) Running awscli tests ... done in 2 minutes and 19 seconds
(6/14) Running healthcheck tests ... done in 0 seconds
(7/14) Running mc tests ... done in 1 minutes and 10 seconds
(8/14) Running minio-dotnet tests ... done in 1 minutes and 42 seconds
(9/14) Running minio-go tests ... FAILED in 4 minutes and 6 seconds
{
  "args": {},
  "duration": 109919,
  "error": "Please reduce your request",
  "function": "testStorageClassMetadataCopyObject()",
  "message": "CopyObject failed on RRS",
  "name": "minio-go: testStorageClassMetadataCopyObject",
  "status": "FAIL"
}

Executed 8 out of 14 tests successfully.
```